Minor Adjustments
I have had to make minor adjustments to the middle maze wall as at some
of the corners the hero was still able to move in and out (due to
canvas inconsistencies such as wall width and hero size). I had to
compensate for all those things. Now, however the hero is completely
restricted by all the walls, corners, etc…